feat: GQL folder structure formatting (#2)
Improving the GQL folder structure for scalability/future-proofing
