--- src/graphql/mutations/user.mutations.ts
@@ -1,7 +1,0 @@
-import { gql } from 'apollo-server';
-
-export const userMutations = gql`
-  type Mutation {
-    createUser(email: String!, name: String): User!
-  }
-`;
