--- src/graphql/resolvers/index.ts
@@ -1,5 +1,0 @@
-import { mergeResolvers } from '@graphql-tools/merge';
-
-import { userResolvers } from './user.resolver';
-
-export const resolvers = mergeResolvers([userResolvers]);
