+++ src/graphql/user/index.ts
@@ -1,0 +1,3 @@
+export * from './user.mutation';
+export * from './user.queries';
+export * from './user.typeDefs';
